Update cli to use HealthResponse type
Moved HealthResponse to its own type definition.
