+++ cli/client/health.go
@@ -9,11 +9,11 @@ import (
 
 const healthPath = "/health"
 
-func (c *RestClient) Health() ([]*types.ClusterHealth, error) {
+func (c *RestClient) Health() (*types.HealthResponse, error) {
 	res, err := c.R().Get(healthPath)
 	if err != nil {
 		return nil, fmt.Errorf("GET %q: %s", healthPath, err)
 	}
-	var healthResponse []*types.ClusterHealth
+	var healthResponse *types.HealthResponse
 	return healthResponse, json.Unmarshal(res.Body(), &healthResponse)
 }
