+++ cli/client/testing/mock_health_client.go
@@ -2,7 +2,7 @@ package testing
 
 import "github.com/sensu/sensu-go/types"
 
-func (c *MockClient) Health() ([]*types.ClusterHealth, error) {
+func (c *MockClient) Health() (*types.HealthResponse, error) {
 	args := c.Called()
-	return args.Get(0).([]*types.ClusterHealth), args.Error(1)
+	return args.Get(0).(*types.HealthResponse), args.Error(1)
 }
